base-files: Add necessary preinit magic so that procd works
Procd was failing to run preinit for three reasons:

Missing initialization file /lib/preinit/00_preinit.conf,
use of netmsg busybox applet that does not exist in openembedded,
and Openwrt uses a substituation of %PATH% with a common default
PATH in a number of locations.

We fix these issues with this patch.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>