basefiles busybox netifd: Allow openwrt-base-image to use netifd without bridge module
in base-files config_generate creates an /etc/config/network which forces
lan to type bridge, and OE also ships a hardcoded network config that did
have type bridge on lan.  Dito netifd for the hardcoded bridged lan.

Bridged lan doesn't work out-of-the-box at the moment because bridge module
is not autoloaded (in Openwrt bridging is built into the kernel so this
doesn't become an issue for them).  Therefore for default config we avoid
bridged configs and just have a straight interface config.

This allows the system to boot with the default static IP address,and
it could also be arranged to be a DHCP client (config changes would
be needed).

Busybox by default has ifupdown enabled which conflicts with netifd's
usage of iup (and no longer has ifdown but executing an extraneous
ifdown can confuse the situation).

We also have netifd conflict with ifupdown package for the same
reason.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>