images packagegroups {base,full}: Fix nomenclature and split
Split the openwrt-image-base into a real 'base' image
and a 'full' image.  'base' is miniimal core (but with GUI),
and 'full' contains more packages.  This could obviously be
even further divided.

Likewise the packagegroups with the addition factor of making
core-boot-openwrt just core-boot modification so that we can
use core-image-minimal.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>