tree-wide: Include Openwrt license when pulling from openwrt repo
Package pulling scripts and configs from openwrt.git were not
indicating the GPL-2.0+ license on that repo.  Therefore add
the LICENSE file from openwrt.git and update LICENSE with an
'&GPL-2.0+'.  Since this is primarily initscripts and such,
while most of the actual software (like procd, ubus) etc is
under an MIT, BSD, or Apache license, it may be desirable to
replace the scripts with MIT, BSD, or Apache scripts.

Closes https://github.com/kraj/meta-openwrt/issues/43

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>