jsonpath: Break CMake custom command dependency cycle
Ninja complains of dependency cycle due to separate custom commands
(which can end up building in parallel but the second has a dependency
on lemon parser build (the first).

| NOTE: VERBOSE=1 cmake --build oe/build/tmp-musl/work/armv5e-oe-linux-musleabi/jsonpath/git-r0/git --target all -- -j 8
| ninja: error: dependency cycle: contrib/lemon -> contrib/lemon

We therefore update to a single add_custom_command that issues
the lemon build then uses it to generate the parser.c that we want.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>