base-files: Add preferopenwrt and preferoe PACKAGECONFIG
Allow to prefer openwrtism (like it's banner and various
symlinks, default dependencies for the base system, etc),
neutral (openwrt packages are used but with more of an OE
flavour), and preferoe which minimizes the openwrtishness
of the system (e.g. when picking and choosing openwrt
packages).  And of course the option to omit openwrt
base-files entirely by removing 'includeopenwrt' from
the PACKAGECONFIG.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>