netifd: Remove dependency on bridge module
There are kernel functions which require bridging early during boot,
e.g. the DSA framework. So if bridging is always a kernel module this
would force people to use an initial ramdisk for startup to load the
modules at an early point in time. Now, it is in the responsibility
of the user to add the bridging functionality as a kernel module or
compiled into the kernel.

Signed-off-by: Kevin Zhang <kevin.zhang3@cn.bosch.com>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>